Changed collections.deque to multiprocessing.queue
Added multiprocessing.queue to share the global queue between the socket process and the main entropy process. Also used a shared "queue_total" variable to share between the proceses. Instead of iterating over the queue each time for the queue total value, we update the value each time a get or a put occurs.